# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Nyx LogBook System API
  version: 1.0.0
  description: Enriched OpenAPI 3.1 specification generated from actual route implementations.
servers:
- url: /
paths:
...
  /supervisor/competency-assessments:
    get:
      summary: GET /supervisor/competency-assessments
      description: Endpoint implemented in src\app\api\supervisor\competency-assessments\route.ts.
      tags:
      - Supervisor
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: &id115
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
        '201':
          description: Error response
          content:
            application/json:
              schema: &id116
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
        '401':
          description: Error response
          content:
            application/json:
              schema: &id117
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
      security:
      - nextAuthSession: []
      x-roles: &id118
      - SUPERVISOR
    post:
      summary: POST /supervisor/competency-assessments
      description: Endpoint implemented in src\app\api\supervisor\competency-assessments\route.ts.
      tags:
      - Supervisor
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: *id115
        '201':
          description: Error response
          content:
            application/json:
              schema: *id116
        '401':
          description: Error response
          content:
            application/json:
              schema: *id117
      security:
      - nextAuthSession: []
      x-roles: *id118
...
  /supervisor/profile:
    get:
      summary: GET /supervisor/profile
      description: Endpoint implemented in src\app\api\supervisor\profile\route.ts.
      tags:
      - Supervisor
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: &id133
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
        '401':
          description: Error response
          content:
            application/json:
              schema: &id134
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
      security:
      - nextAuthSession: []
      x-roles: &id135
      - SUPERVISOR
    put:
      summary: PUT /supervisor/profile
      description: Endpoint implemented in src\app\api\supervisor\profile\route.ts.
      tags:
      - Supervisor
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: *id133
        '401':
          description: Error response
          content:
            application/json:
              schema: *id134
      security:
      - nextAuthSession: []
      x-roles: *id135
...
  /supervisor/review:
    get:
      summary: GET /supervisor/review
      description: Endpoint implemented in src\app\api\supervisor\review\route.ts.
      tags:
      - Supervisor
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
        '401':
          description: Error response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
      security:
      - nextAuthSession: []
      x-roles:
      - SUPERVISOR
...
  /worker/review:
    get:
      summary: GET /worker/review
      description: Endpoint implemented in src\app\api\worker\review\route.ts.
      tags:
      - Worker
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: &id160
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
        '401':
          description: Error response
          content:
            application/json:
              schema: &id161
                type: object
                additionalProperties: true
                description: 'TODO: infer response object schema'
      security:
      - nextAuthSession: []
      x-roles: &id162
      - SUPERVISOR
    post:
      summary: POST /worker/review
      description: Endpoint implemented in src\app\api\worker\review\route.ts.
      tags:
      - Worker
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: *id160
        '401':
          description: Error response
          content:
            application/json:
              schema: *id161
      security:
      - nextAuthSession: []
      x-roles: *id162
components:
  securitySchemes:
    nextAuthSession:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie used for authenticated requests.